Fix sign-compare

+++ src/reference_generator.cpp
@@ -432,7 +432,7 @@ bool ReferenceGenerator::calculatePositionReferencesInternal(JointGoal& goal, do
         ++goal.sub_goal_idx;
 
         // Check if this was the last trajectory point. If so, this goal is finished!
-        if (goal.sub_goal_idx >= goal.goal_msg.trajectory.points.size())
+        if (goal.sub_goal_idx >= static_cast<long>(goal.goal_msg.trajectory.points.size()))
         {
             for(unsigned int i = 0; i < goal.num_goal_joints; ++i)
                 joint_info_[goal.joint_index_mapping[i]].goal_id.clear();
@@ -470,7 +470,7 @@ bool ReferenceGenerator::calculatePositionReferencesInternal(JointGoal& goal, do
             {
                 sub_goal_velocities = sub_goal.velocities;
             }
-            else if (goal.sub_goal_idx + 1 < goal.goal_msg.trajectory.points.size())
+            else if (goal.sub_goal_idx + 1 < static_cast<long>(goal.goal_msg.trajectory.points.size()))
             {
                 for(unsigned int i = 0; i < goal.num_goal_joints; ++i)
                 {
